[BUG] Fix potential rewards stuck for AaveV3SCY
This bug is rather interesting. AaveV3 will remove a token from its
reward list once the token is no longer given out. This, however, breaks
all the accounting of the RewardManager.

To mitigate this, we need a way to allow adding new reward tokens to the
list & at the same time not allow the removal of already given out
tokens. Having a owner-only function doesn't seem like the best idea,
hence we have implemented a permission-less function to allow this,
please check out `addNewRewardToken`